% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citysummary.R
\name{citysummary}
\alias{citysummary}
\title{Descriptive statistics of data}
\usage{
citysummary(x = simpledata)
}
\arguments{
\item{x}{the dataset}
}
\value{
a data.frame
}
\description{
This function presents the descriptive statistics information of the dataset
}
\examples{
citysummary(simpledata)
}
