% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costsummary.R
\name{costsummary}
\alias{costsummary}
\title{Brief Summary of Costs}
\usage{
costsummary(x = "Beijing")
}
\arguments{
\item{x}{the city name}
}
\value{
a data.frame
}
\description{
This function collects the brief summary and tidy it into a table.
This function obtains the index data from Numbeo (https://www.numbeo.com/)
You could input any city name.
}
\examples{
costsummary("Beijing")
}
