% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbeo.R
\name{numbeo}
\alias{numbeo}
\title{Data from Numbeo}
\usage{
numbeo(x = "cost")
}
\arguments{
\item{x}{the category, including "cost", "crime", "quality".}
}
\value{
a data.frame
}
\description{
This function obtains the index data from Numbeo (https://www.numbeo.com/)
You could input 'cost', "crime", or 'quality'.
}
\examples{
numbeo("cost")
}
