% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{quality}
\alias{quality}
\title{Presenting Quality Indices}
\usage{
quality(x = "all")
}
\arguments{
\item{x}{the city, a city name or "all"}
}
\value{
a list include a table and a plot
}
\description{
This function presents the life quality indices.
}
\examples{
quality("Beijing")
}
