% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Presenting scores of city}
\usage{
score(x = "all")
}
\arguments{
\item{x}{the city, input a city name or "all"}
}
\value{
a plot
}
\description{
This function presents the average score of each city.
}
\examples{
score("Tokyo")
}
