% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleport.R
\name{teleport}
\alias{teleport}
\title{Index from teleport}
\usage{
teleport(x = "city")
}
\arguments{
\item{x}{the category}
}
\value{
a data.frame
}
\description{
This function obtains the index data from teleport (https://teleport.org/)
You could input 'city', any city name, or 'all'.
}
\examples{
teleport("Beijing")
}
